centralize default empty matrices
