+++ antarest/study/storage/rawstudy/model/filesystem/common/prepro.py
@@ -9,14 +9,18 @@
 # SPDX-License-Identifier: MPL-2.0
 #
 # This file is part of the Antares project.
-import numpy as np
 from typing_extensions import override
 
 from antarest.study.storage.rawstudy.model.filesystem.config.model import FileStudyTreeConfig
 from antarest.study.storage.rawstudy.model.filesystem.folder_node import FolderNode
 from antarest.study.storage.rawstudy.model.filesystem.ini_file_node import IniFileNode
 from antarest.study.storage.rawstudy.model.filesystem.inode import TREE
-from antarest.study.storage.rawstudy.model.filesystem.matrix.constants import default_scenario_hourly
+from antarest.study.storage.rawstudy.model.filesystem.matrix.constants import (
+    default_conversion,
+    default_data,
+    default_k,
+    default_scenario_hourly,
+)
 from antarest.study.storage.rawstudy.model.filesystem.matrix.input_series_matrix import InputSeriesMatrix
 
 
@@ -28,20 +32,6 @@ class PreproCorrelation(IniFileNode):
 class PreproAreaSettings(IniFileNode):
     def __init__(self, config: FileStudyTreeConfig):
         IniFileNode.__init__(self, config)
-
-
-def default_k() -> np.ndarray:
-    return np.zeros((24, 12), dtype=np.float64)
-
-
-def default_conversion() -> np.ndarray:
-    return np.array([[-9999999980506447872, 0, -9999999980506447872], [0, 0, 0]], dtype=np.float64)
-
-
-def default_data() -> np.ndarray:
-    res = np.ones((12, 6), dtype=np.float64)
-    res[:, 2] = 0
-    return res
 
 
 class PreproArea(FolderNode):
