+++ antarest/study/storage/rawstudy/model/filesystem/matrix/constants.py
@@ -39,3 +39,45 @@ def default_6_fixed_hourly() -> np.ndarray:
 
 def default_8_fixed_hourly() -> np.ndarray:
     return np.zeros((8760, 8), dtype=np.float64)
+
+
+def default_k() -> np.ndarray:
+    return np.zeros((24, 12), dtype=np.float64)
+
+
+def default_conversion() -> np.ndarray:
+    return np.array([[-9999999980506447872, 0, -9999999980506447872], [0, 0, 0]], dtype=np.float64)
+
+
+def default_data() -> np.ndarray:
+    res = np.ones((12, 6), dtype=np.float64)
+    res[:, 2] = 0
+    return res
+
+
+def default_maxpower() -> np.ndarray:
+    return np.zeros((365, 4), dtype=np.float64)
+
+
+def default_reservoir() -> np.ndarray:
+    return np.zeros((365, 3), dtype=np.float64)
+
+
+def default_credit_modulation() -> np.ndarray:
+    return np.zeros((2, 101), dtype=np.float64)
+
+
+def default_water_values() -> np.ndarray:
+    return np.zeros((365, 101), dtype=np.float64)
+
+
+def default_energy() -> np.ndarray:
+    return np.zeros((12, 5), dtype=np.float64)
+
+
+def default_link_legacy_matrix() -> np.ndarray:
+    return np.zeros((8760, 8), dtype=np.float64)
+
+
+def cost_level() -> np.ndarray:
+    return np.full((8760, 1), -1e-6)
