+++ antarest/study/storage/rawstudy/model/filesystem/root/input/hydro/common/capacity/capacity.py
@@ -20,7 +20,13 @@ from typing_extensions import override
 from antarest.study.model import STUDY_VERSION_6_5, MatrixFrequency
 from antarest.study.storage.rawstudy.model.filesystem.folder_node import FolderNode
 from antarest.study.storage.rawstudy.model.filesystem.inode import TREE
-from antarest.study.storage.rawstudy.model.filesystem.matrix.constants import default_scenario_daily
+from antarest.study.storage.rawstudy.model.filesystem.matrix.constants import (
+    default_credit_modulation,
+    default_maxpower,
+    default_reservoir,
+    default_scenario_daily,
+    default_water_values,
+)
 from antarest.study.storage.rawstudy.model.filesystem.matrix.input_series_matrix import InputSeriesMatrix
 
 
@@ -29,22 +35,6 @@ class MatrixInfo(TypedDict, total=False):
     freq: MatrixFrequency
     start_version: StudyVersion
     default_empty: Callable[[], npt.NDArray[np.float64]]
-
-
-def default_maxpower() -> npt.NDArray[np.float64]:
-    return np.zeros((365, 4), dtype=np.float64)
-
-
-def default_reservoir() -> npt.NDArray[np.float64]:
-    return np.zeros((365, 3), dtype=np.float64)
-
-
-def default_credit_modulation() -> npt.NDArray[np.float64]:
-    return np.zeros((2, 101), dtype=np.float64)
-
-
-def default_water_values() -> npt.NDArray[np.float64]:
-    return np.zeros((365, 101), dtype=np.float64)
 
 
 INITIAL_VERSION = StudyVersion.parse(0)
