+++ antarest/study/storage/rawstudy/model/filesystem/root/input/link/area/area.py
@@ -9,7 +9,6 @@
 # SPDX-License-Identifier: MPL-2.0
 #
 # This file is part of the Antares project.
-import numpy as np
 from typing_extensions import override
 
 from antarest.matrixstore.matrix_uri_mapper import MatrixUriMapper
@@ -17,16 +16,15 @@ from antarest.study.model import STUDY_VERSION_8_2
 from antarest.study.storage.rawstudy.model.filesystem.config.model import FileStudyTreeConfig
 from antarest.study.storage.rawstudy.model.filesystem.folder_node import FolderNode
 from antarest.study.storage.rawstudy.model.filesystem.inode import TREE
-from antarest.study.storage.rawstudy.model.filesystem.matrix.constants import default_6_fixed_hourly
+from antarest.study.storage.rawstudy.model.filesystem.matrix.constants import (
+    default_6_fixed_hourly,
+    default_link_legacy_matrix,
+)
 from antarest.study.storage.rawstudy.model.filesystem.matrix.input_series_matrix import InputSeriesMatrix
 from antarest.study.storage.rawstudy.model.filesystem.root.input.link.area.capacities.capacities import (
     InputLinkAreaCapacities,
 )
 from antarest.study.storage.rawstudy.model.filesystem.root.input.link.area.properties import InputLinkAreaProperties
-
-
-def default_link_legacy_matrix() -> np.ndarray:
-    return np.zeros((8760, 8), dtype=np.float64)
 
 
 class InputLinkArea(FolderNode):
