+++ antarest/study/storage/rawstudy/model/filesystem/root/input/st_storage/series/area/st_storage/st_storage.py
@@ -9,21 +9,17 @@
 # SPDX-License-Identifier: MPL-2.0
 #
 # This file is part of the Antares project.
-import numpy as np
 from typing_extensions import override
 
 from antarest.study.model import STUDY_VERSION_9_2
 from antarest.study.storage.rawstudy.model.filesystem.folder_node import FolderNode
 from antarest.study.storage.rawstudy.model.filesystem.inode import TREE
 from antarest.study.storage.rawstudy.model.filesystem.matrix.constants import (
+    cost_level,
     default_scenario_hourly,
     default_scenario_hourly_ones,
 )
 from antarest.study.storage.rawstudy.model.filesystem.matrix.input_series_matrix import InputSeriesMatrix
-
-
-def cost_level() -> np.ndarray:
-    return np.full((8760, 1), -1e-6)
 
 
 class InputSTStorageAreaStorage(FolderNode):
